Fixed some instruction translation issues
Right now, I think instructions actually translate correctly for
conditional jumps, and unconditional jumps with immediate values.
For the latter, I have implemented the special case for callbacks
being passed to library functions resolved dynamically using the
plt.  However, I also realized that the current approach for the
brute-force disassembly technique does not work; the first byte
of an instruction can't simply be replaced with its translation.
This is a problem because immediately after that transformed
instruction, if the remaining bytes of the instruction stay, they
are now garbage bytes in this control flow (although starting
from a different offset, they may be valid).  Therefore, somehow
we need to efficiently store valid disassemblies for different
offsets, and somehow be able to transition from one to another
by using the mapping (a jump could switch to a different valid
disassembly of the same code).  This requires more thought.