Started progress on correct mapping
The mapping phase itself should now correctly create a mapping.
However, we may still want to put jumps at the end of each
block of code (if it doesn't end in a jmp or ret) so that control
can flow to another block at another location in the new code.