Added instruction translation for REG and MEM operands
Now all instructions should in theory be correctly translated
for relocating the text of the binary.  However, the lookup
function for the mapping and the ability to write the mapping
itself to a low-level hash table (the one that will be embedded
in the binary itself) have not been implemented yet.

Also added was another assembly example to test the translation
of jmp/calls to an address stored in memory or a register such as
jmp eax.