Added mapping output function and created lookup prototype
Outputting the mapping was surprisingly simple.  After analyzing
Brian's solution, I realized that it was not necessary to use
a paged solution, as it has limited benefit in a dense mapping.

Instead, I simply output the mapping of every byte in the original
binary's executable segment, placing 0xffffffff in the case of
any unmapped byte.  The raw bytes are output in little-endian,
but other than that, there is no data structure at all; just
essentially an array of 4-byte addresses that can be indexed by
subtracting the base address from the original address, thus meaning
that (assuming offset = original-base) the mapping can be retrived
by reading from [mapbase + offset*4].

I also produced a prototype for the mapping lookup function in
lookup.asm, which has the mapping from the "mem" program stored
in the data section.  I modified renabler.py to output the mapping
in a format accepted by nasm, and pasted the mapping in.

lookup.asm calls the prototype lookup function twice, attempting
to retrieve and print two different addresses from the mapping.
The lookup function has some hard-coded values that will be
custom to each rewritten binary and therefore will become "holes"
in the lookup template in renabler.  These are:
-base of original executable segment
-size of original executable segment
-offset of the instruction that retrieves eip in lookup from the
	start of the new executable code (in the prototype, it is
	just an offset that makes the final printed values easy
	to understand).