Added lookup function and placed working mapping in new text
The lookup function correctly performs a lookup of an address
at runtime.  However, differently than before, if the address is
outside the mapping's range, instead of halting, the lookup
function instead returns the original address, allowing jmps/calls
outside of the original text section (as would be the case with
dynamically loaded libraries, which operate in a different address
region than the binary's text.

The mapping is now complete for all jmp/call instructions and
almost all conditional jumps (still missing jcxz and jecxz).
In addition, the remapped assembly programs appear to work
flawlessly, although I have only tested a couple of them so far.

The C program, simplest, however, segfaults before it reaches the
main function.  After a lot of debugging, I have tracked the
problem down to __x86.get_pc_thunk.bx, which I believe is the
source of the issues.  The problem appears to be that _start
uses the program counter to compute the location of .got.plt, and
then subtracts 4 bytes from that to get what I assume is the last
entry in .got.

Our modifications break this, because the program counter is in
a different place, so the data retrieved from the computed address
will be nonsense.  I figure if I use a special case to check for
calls to this function, I may be able to avoid this issue by
patching the address (a reverse lookup).  This is basically
dealing with position independent code.