Added __x86.get_pc_thunk.bx special case workaround
Whenever a function calls __x86.get_pc_thunk.bx, I patch the
call to replace the call with a mov instruction that sets
ebx to what the function would have returned in the original code.
This should be safe because if the thunk is ever used to
obtain an address and then makes a jmp/call to the code at that
address, the call itself will be patched with a call to the
dynamic symbol table.

However, I noticed that at the end of execution, in the call to
exit() (called at the end of __libc_start_main).  exit() calls
a function, __call_tls_dtors, which somehow obtains the address
to __do_global_dtors_aux, thereby executing code in the old
text section at the very end of execution.

That said, programs seem to run successfully.  I tested with
simplest, which printed the correct output and exited correctly,
and then I tested bzip2, which took a very long time to process.
The modified binary successfully compresses and decompresses
a test file, but one strange error is that it rejects certain
input flags that the original binary supports just fine.
For example, for this action of the original binary:
./bzip2 -d compress_me.txt.bz2
The first of these works, and the second doesn't:
./relocated --decompress compress_me.txt.bz2
./relocated -d compress_me.txt.bz2
The latter invocation gives an error message about a
"Bad flag `X-�'".  I will need to investigate this further.