Added assembler cache and other optimizations
The speed has increased massively; for simplest, time decreased
from 40-50 seconds to 3 seconds, and for bzip2, from upwards of
half an hour down to 57 seconds.

I performed 3 main optimizations.  The first was to add a simple
cache for the assembler, simply matching the exact text of the
passed assembly with the cache.  This increased speed a lot, but
certain assembly used unique values for offsets, making the cache
less useful.

Therefore, the second optimization was to transform all
jumps with offsets to use a dummy value, and then patch the
resulting binary with the real address.  This gave another large
speedup. (TODO: Double-check that no non-control-flow instructions
got clobbered by this)

The final optimization involved the brute_force_disasm function,
which ended up spending most of its time iterating through each
submapping.  I instead created a dummy mapping only used for
checking whether we already had encountered that address, making
checking whether we had encountered an address an O(1) operation.

I did note that the output from lookup is incorrect, which is
very unusual, so I need to check what is wrong there.