Added rewriting of ret, changed rewriting of call
After a discussion, it was decided that position independent
code can be handled simply by changing the way we patch the
call instruction.  Call is special because it pushes the
instruction pointer on the stack and is the only way to access
the instruction pointer in 32-bit x86.  Therefore, if we replace
all call instructions with a push instruction (pushing the
original return address) and a jump, and all ret instructions
with a pop of the old address, a lookup of the old address, and
a jmp to the mapped address.

A side-effect of this rewriting is that as soon as a call is made
to an address outside the relocated text, the return address
on the stack points to the old text section, and therefore
execution will continue in the old text section after that
call.  Calls within the new section are rewritten to perform the
lookup and still work.  What this means is that all code for a
binary will need to be rewritten, including dynamic libraries.
This will require the ability to rewrite .so files.

Other small edits in this commit were fixing msearch.py to not
raise an exception when an index is not found, commenting out
many of the debug statements in renabler.py, and adding .pyc
files to the .gitignore.