Continued optimizations
Assembler runs faster now, but I would like to more thoroughly
test its correctness, as well as the correctness of the
new transformations for call and ret.

I need to implement ret <immediate> instructions,
as well as some exotic ret instructions that tend to appear
when using brute force disassembly.  For now I ignore them.
A potential temporary measure could be replacing them with
hlt so I can find exactly where a problem occurs if one of the
unimplemented instructions is actually encountered.

I also tried to test a statically linked binary, but it turns
out the kernel does some tricky dynamic linking anyway and
therefore control exits from the new code before the _init
function (actully exits in a call to uname, when it makes
a system call), so I was unable to test very much of the
patched application.

We intend to support dynamic linking, of course, so soon
we need to start being able to patch .so files, but it
appears we may run into several complications.