Added test assembly for looking up auxiliary vector
The AT_SYSINFO entry in the auxiliary vector holds the
code used to perform a syscall in libc (if the kernel provides it).
This special case ends up returning to the wrong address
if we put old addresses on the stack, so if there is a call
to the address in this entry we need to replace the old return
address on the stack with the mapped return address so it returns
to the correct address.  I wrote a simple assembly function to
test retrieving the address for this special case, which I think
will have to run before the application's new entry point and
save the address somewhere for the global lookup function.