Added auxv scanning, fixed saving vals bug, fixed bug in assembler
Now ssimplest runs successfully, and it appears that execution
never leaves code we have modified save for the code that
performs the sysenter.  For that, the special case is now working:
at the start of the program, I scan the stack to find the
auxiliary vector, find the AT_SYSINFO entry, and save the address
it holds to a special location.

The special location is currently the end of the new text section;
both this address and the global lookup function currently reside
there so I don't have to make a binary with two segments (due
to current limitations of the elf writer we can't do that for the
static binary).

The global lookup function, as written right now, is a hack.
Since the application I am testing is statically linked, there
is only one module and no .so files other than the vdso.
Therefore, I am using the global lookup function only to overwrite
the stack value of the return address when the address is the
AT_SYSINFO address, which is the special case when we must
update our stack value before making a syscall.

By adding the global lookup function, which calls the local
lookup function again, we end up pushing a number of bytes
on the stack; too many, actually, as it overwrote the values
I had been saving outside the stack.  I moved the values further
outside the stack, but the real global lookup function may force
me to move them even further.  I have noticed that this is a
source of bugs, so it's something to watch.

In addition, the optimizations for the assembler ended up having
a disastrous effect when one of the instructions did not
assemble correctly.  The reason for this is that for an
instruction of the form
mov eax, dword ptr [esp + 0x6f]
the generated instruction, as the optimization did it, was wrong.
It appears that using esp results in an encoding with one more
byte than when using other registers.  This does mean that I need
to be extra careful with the optimizations, but honestly I wish
I could just write a faster assembler that doesn't just use
external programs to assemble.  It seems as if it should be
possible to make assembling much, much faster, but x86 encodings
are unfortunately quite complicated, so writing an assembler would
be a substantial task.