Moved lookup function to text section start, optimized assembler
These changes are towards the general effort of implementing
the global mapping, which needs to somehow know the locations
of local lookups.  The next step is to add the code that
parses /proc/./maps and gives us these mappings.  By putting
the local lookup at the start of each text section, all we
need to know is the start of the address range.

The assembler had slowed down again due to the added code for
ret instructions.  I added additional instruction-specific
caching code.