Added ability to rewrite .so files, added hlt to global lookup
Now, when an entry is invalid in the global lookup, it jmps to a
hlt instruction.

Getting the right .so to load is a bit complicated, as we must
both turn off lazy loading and override the default library
location (as with libc).

It appears that popgm still does not work, or the way I index
the values in popgm does not work, as I can only find invalid
entries around the address that is looked up in my test
application.