Fixed local lookup for .so files, updated popgm
It turns out that the local lookup function did not function for
shared libraries because it had the implicit assumption that
the base address of the static library was the same when it is
loaded; of course, a dynamic library is loaded at a random
base address, so it always considered the lookup address as
outside the range.  In addition, it was not making comparisons
for unsigned integers (it was using jl and jge rather than jb and
jae).

Dynamic libraries still do not work, unfortunately.  Each call
instruction is rewritten to push the old return address on the
stack.  Unfortunately, the old return address is also based on the
implicit assumption that the base is 0; all the addresses do not
match the runtime addresses.  For this, we may need to use PIC for
every single call instruction in an .so.  We can use the
instruction's address to determine the base address of the new
text section, and from that we can determine the base address of
the old text section.  This is similar to the fix I had to make
for the local lookup function, but a little more complicated.
I am, however, worried about the additional overhead this will
incur.

There were also issues with populate_gm, but we managed to fix
them.  It appears to be working correctly now.