Made so calls PIC, added exit for 0xffffffff, added asm optimizations
Calls in an .so follow the formula

	((address - new_offset) - new_base) + old_offset

There are problems with libc and exit; it wants to jump to the
dynamic linker, so I made it simply abort cleanly.

Unfortunately, certain calls also seem to make it want to jump
to the linker, so we will rewrite return values as a future
workaround.