Update to handle calls into dynamic linker, updated popgm, asm
Now populate_gm will fill the range of the dynamic linker's code
pages with zeroes.  If the global lookup encounters zeroes, it
performs a lookup of the return address on the stack and overwrites
it with the new address before execution continues into the linker.

In addition, some minor changes were made to the assembler to
handle some rare edge cases.