Added change to support new gdb commands
I just wrote a python script for gdb adding smap and rmap commands,
which allow for searching for the corresponding new address to
an original address, and a reverse search for the original address
from the new address.

Now whenever a new binary is written, a <binary>-r-map.json file
is written as well, which is the mapping for that specific file.