Fixed reg. corruption bug, null-termination bug, updated ELFManip
One problem that had manifested itself in all binaries was that
they all died at the very end, in the call to exit.  In fact,
it was when it was trying to call _dl_fini when exit called
__run_exit_handlers.  The issue appeared to be caused when the
macro PTR_DEMANGLE was performed to retrieve the pointer to
_dl_fini, which uses a value from the TCB to "decrypt" the pointer.
I tracked the pointer back to where it was first set at the start
of the program in the call to __internal_atexit from __cxa_atexit.
This revealed that the corruption occurred even earlier.  I
finally traced it all the way to the entry point of the program,
at which point I discovered that the call to popgm that we
inserted before _start would not preserve the values in registers.
Since popgm is C, it modified several registers, and I didn't
check whether the registers held any meaningful values.  Thus,
by saving all registers, this problem was finally solved.

However, there ended up also being a problem in the popgm source
as well.  The code null-terminated the end of the buffer used for
reading the memory maps, but did not null-terminate the actual
bytes read into the buffer, leaving a part of the buffer filled
with garbage bytes.

I also updated ELFManip.  The first update did not work immediately
but after a quick update by Brian, everything seems to work well.
I will try to do more testing to make sure everything except for
multithreading works.  Incidentally, the reason coreutil's sort
executable causes a segfault is almost certainly because
it spawns multiple threads and trips the global lookup's multiple
call guard.