Changed global lookup flag to be thread-specific
Now when rewriting a binary a new section will be added that
includes the byte for the global lookup flag, which will be
loaded as part of the thread local storage for each thread.

As of right now, this results in a TLS section at address
0x06ff0000, which violates an assumption made in populate_gm.c.
However, the code still works correctly; I believe that popgm
adds a spurious entry to the global mapping, but that it will
never be read and therefore is not a concern for now.

However, if the original TLS data is larger than 0x10000 bytes,
this will fail.  Eventually we need to be able to place this at
an arbitrary address, but for now it works.