Fixed assembly bug
I previously found that, when trying to save addresses that the
program was jumping to in eax, that if the original instruction
used only the lower half of the register (e.g. ax or bx), pwntools
would fail with a fatal warning due to the register sizes
mismatching.  Therefore I would change to the full register
(e.g. change bx to ebx).  One of the libraries I was rewriting
used di instead of edi, and I had matched only on registers ending
with x, so I added support for registers ending in x or i.