Fixed popgm bug
As we were worried about previously, just because you request
n bytes in a read syscall, you won't necessarily get that many.
Therefore, when trying to read from /proc/self/maps, we had
no guarantee that we were in fact reading the entirety of the
memory mappings.  This was not a problem for programs with only
a few libraries, but once it expanded to more than a handful, the
later ones were not read and were therefore not mapped, leading
to a crash.  Now it attempts to read the entire file.

The buffer to hold the file contents is now 0x10000 bytes
(from 0x1000), so if /proc/self/maps is larger than that it
will still fail.