Partially fixed rare code alignment issue
If an .so file has a call instruction near the very start, the
placeholder value used in a sub instruction will result in the
long encoding of that instruction, while the final offset will
result in the short encoding of that instruction, leading to
an incorrect code size and broken offset to the lookup mapping.

This is fixed by doing only one subtraction and combining the two
values that were subtracted previously.  This would still cause
issues if that resulting value could be small, but this is
unlikely because it includes the offset that the code must be from
the original text section (which must be after the text, data, etc.
sections).

However, I have just noticed that I applied this fix only for
unconditional calls, so this will also need to be done for
conditional call instructions as well.