Added rewrite.py
This python script is a convenience, automatically rewriting
all the libraries for a binary and placing them in a directory
by the name of "<binary>-libs-r".  It also rewrites the binary
itself, making the act of rewriting the binary a single, simple
fire-and-forget task.

However, if a library has already been rewritten for another
binary, the script has no way of knowing this and skipping the
time-consuming and unnecessary rewriting process for that library.
I would eventually like to add some common area for already
rewritten libraries that we can simply copy instead of rewriting.