Removed excessive padding in .so files
The offset in .so files for the new text section was 0x9000000,
just like the new text section's address in the main binary.
This resulted in each library taking a huge amount of space
in memory, and as a result binaries with a large number of
libraries crashed on startup due to running out of address space.

Now the offset in .so files of the new text section is on
the next page after the highest address of all the existing
sections (highest address calculated by finding the segment with
the largest segment address + size).