Moved popgm to global mapping section
Now popgm is placed directly after the global lookup.  This will
be useful for popgm to be called from a library.

In addition, I added a small assembly snippet to the front of
popgm that saves and restores the current register state so that
any code calling popgm doesn't need to worry about popgm clobbering
its state.