Added code to handle dynamically loaded libraries to rewrite.py
This is an old change that was not committed at the time I
finished it.  If I recall correctly, a file listing all the
dynamically loaded libraries (loaded after program start, e.g.
with dlopen) is required so that those libraries can also be
rewritten and passed to the dynamic linker so that it (hopefully)
chooses those libraries instead of the originals.  However, this
still does not work for (I think, but perhaps it might not work for
other things too) loaded libraries that  have a fixed path to
the library embedded in the application.  This discovery was the
point that I started feeling existential despair, because it
seems almost impossible to create a general way to guarantee the
application will definitely use the rewritten libraries without
replacing the originals...