Added preliminary support for flag to only rewrite main binary
Added -execonly flag, which enables a different lookup function
to be added to the new text section.  This lookup function
does not use the global mapping at all (and in fact the global
mapping will not be needed and we can get rid of it in a binary
rewritten with this flag, as well as popgm, and auxvec.

However, rewriting only the executable re-introduces some
challenges, such as the fact that we must again concern ourselves
with callbacks in functions that are executed in libraries but
call functions in the main executable.  I have created a prototype
function that currently only identifies when there is a library
call that *does* have callbacks by checking PLT entries.  Each
entry, as extracted by pyelftools, tells us the corresponding
GOT entry that will hold the destination address after the dynamic
linker resolves the function's address.  Therefore, we can check
the destination address of each jmp (and call to be safe, but I
think it should never happen) for a destination in the GOT.  This
will tell us that we are in fact in the PLT entry for that
library function and that we can rewrite the callback arguments
on the stack.  This does assume that the only way we access such
functions is through the PLT.  If that assumption is violated,
this method *will* fail.

However, it still needs to rewrite all of the
function addresses pushed on the stack as arguments.  In addition,
there is only one function that is identified as having callbacks,
__libc_start_main (which takes 3 function addresses), but we need
to have many.  These must be somehow determined from libraries,
possibly manually by analyzing the API, and therefore this is
a sticking point for automation.  However, once it's added we can
always check for such a function in the PLT and therefore once
we have a sizable library of common library functions with
callbacks, encountering a new function that takes callbacks should
probably be rare.