Added flag and placeholders for removing support for generic PIC
When removing generic PIC, we will no longer push old addresses
on the stack and translate those old addresses to new addresses
with each rewritten ret instruction.  Instead we push new addresses
with call instructions and use the original ret instructions.  This
should result in better performance because we will not be
obfuscating the call stack any more.