Finished implementing -nopic flag
Now, when support for generic PIC is removed, there does indeed
appear to be a performance boost.  We don't rewrite calls and
returns to push/jmp and pop/jmp combinations with this flag,
reducing the number of instructions executed.  There is one
exception to this: __x86.get_pc_thunk.bx.  This appears to be
in a binary even when it is not compiled for PIC (I think because
we are including standard libraries that are PIC, and they use
it).  Therefore, for calls to that specific function, and for the
ret instruction in that function, we DO push the old address on
the stack and pop/translate it on return.

What's also nice is that we don't have to rewrite return addresses
for when you exit the main binary!  I use the secondary lookup
for this since it doesn't rewrite addresses on the stack.

One caveat right now is that I assume you also are only writing
the main executable for this flag.  I haven't really explored
the feasibility of this optimization when also rewriting shared
objects.  This may be something to eventually explore.  Might be
interesting.