Added working example instruction counting instrumentation script
The file icount.py should count all the original program
instructions that executed.  It currently recycles the
global_sysinfo variable, since I was testing with the -execonly
flag, and therefore that field wasn't used.  I used a technique
discussed in another paper that is supposed to be very efficient,
but it is a little counterintuitive.  It uses the loop instruction
because it doesn't set flags, but loop decrements.  Therefore, we
count *down* from 0xffffffff and get the final number of
instructions that were executed by subtracting the final value
from 0xffffffff.  However, when only rewriting a main binary, I'm
not sure how to insert code right before the program ends in
order to determine the number of instructions.  I manually set
a breakpoint at the last instruction of the main function in a
simple example program, and I could check the value of the counter
in gdb, but for more complex programs, the main function may have
multiple return locations!  If I'm ALSO using -nopic, I can't
even analyze the return address (at least not without inserting
more instructions just for this purpose).  It could be possible
to insert instructions (from the user side) for every ret,
checking if it is a return from main into __libc_start_main, but
I'm not sure how inefficient that would be.  Alternatively, I
could use the address ranges for the main binary to only
instrument ret instructions inside the main function.  That
could be promising...

I also made some changes for the translation of conditional jumps.
What is interesting is that I did not account for the offset of
instructions inserted before such instructions until now.  For
most jcc instructions, it didn't matter because there wasn't
anything inserted before them. but that wasn't the case with
jcxz and jecxz.  It inserted one instruction before the replacement
instruction (jz), but the new target was still computed based
on the assumption that the jz started at an offset of zero for
the rewritten instruction.  I don't understand how this ever worked
if there were ever jcxz/jecxz instructions actually used in any
of the programs I tested, because the destination would be
slightly wrong.  I changed it to compute the correct offset of
the jump instruction, shifted by the single inserted test
instruction and any user instrumentation instructions, and things
seem to work fine.  I just don't get how that didn't cause a
glitch sometime earlier.