Started refactoring code to make project more modular, less fragile
Hopefully this will make different pieces of the project easier
to work with, as well as make it easier to understand for somebody
else who might need to work with it in the future.

The brute force disassembler has been separated out successfully,
and I think is a bit less tightly coupled than before.  I still
rely on it yielding None when it reaches the end of a sequence,
which is why I designed the brute force mapper to work closely
with it instead of having a generic mapper that isn't aware whether
or not we are doing brute force disassembly.  Since the mapping
process has some significant differences when we are using brute
force disassembly, I figured it was fine to have a corresponding
mapper to deal with the mapping-level issues surrounding it.

mapper.py and brute_force_mapper.py are not actually being used
yet, because they really need access to a rewriter object before
they can be completed.  Therefore, for now the code I adapted for
the mapper was pasted in the main binary with a small amount of
bootstrapping code (and self references removed) for testing.
The modified mapper code works with the new disassembler, so now
it is time to create the more complex rewriter/translator
object and then create an instance of it in the mapper.
Then we will have the major components separated out, and THEN
I can start creating an alternate rewriter/translator for x86-64.