Added translator object and context object
The translator will be used to translate all instructions
from an original binary to their corresponding instructions in
a rewritten binary.  This refactoring of code will allow for
(hopefully) easier modification of the translation to support
both x86 and x86-64 binaries without adding too much complexity.

The context object is a collection of what used to be global
variables, now collected into one object.  I'm not sure about
whether this is really the best design decision, but I have not
yet thought of a better way to organize the assorted global
variables.