Added runtime.py and x86_runtime.py, continued with refactoring
The new runtime files were added to provide the functions needed
for rewritten binaries to function, and thus these functions
comprise the runtime libraries for the system.  Most of the code
for these is written in assembly, but popgm is in C, and in theory
different implementations could be written in different ways.  I
added this first so that I could better organize the code, but
also because this code is platform dependent (i.e. I will have
to write new assembly for each platform supported).