Tested rewriting with all flags, renamed assembler
The assembler will need to be platform specific, so I moved
assembler.py to x86_assembler.py.  Right now, it doesn't use an
object, but it could be changed to be more in line with the other
platform-specific files I have made.  I am, however, thinking of
switching to a different assembler completely: keystone.