Added initial x86-64 support, still broken, changed args parsing
I have switched to using the argparse library for the arguments
so that I don't have to deal with the growing number of arguments
to renabler.py.  I did this now because I have added the --arch
flag, allowing selection of the architecture.

Right now, x64 support doesn't actually work.  The first problem
I have encountered with it so far is that pop/push instructions
that I wrote for 32-bit code can't be used; not only would the
meaning be different since addresses are a different size, such
instructions are not even allowed on the stack.  However,
worryingly, it seems as if pwntools does not like the 64-bit
registers, as it seems to think they are relocation entries, even
using the 'amd64' architecture for pwnlib.  If I can't get pwntools
to work with 64-bit assembly, I will have to use keystone instead,
which does seem to work.

Strangely, pwntools will assemble something like "pop eax", but
not the *first* time you try---the first time it says there is
an "operand type mismatch", but the second time it produces the
byte "58", which I think would be valid in x86, but I'm not sure
if it's valid x86-64.  This could use further investigation, but
the bottom line is that I may not be able to use pwntools for
x86-64.