Modified x64_Translator to start to support x86-64
There are a lot of small changes that need to be made, and even
then, I'm not sure about whether things will work.  There's a lot
to consider here.  I did a bit of research into accessing RIP
before I started making changes here to make sure I was handling
it in the right way.  I will need to make a lot of changes with
rewriting PIC code, specifically code that references RIP.  This
will mean somehow calculating the intended offset from RIP in
a shared object.  Since we won't know the base address in a shared
object, all calculations must be relative.  I was thinking that
in order to get the right offset I need to know the offset between
the old text section and new text section statically, which should
make it possible to generate not-too-expensive runtime instructions
by doing the calculations offline.  Something like converting the
new RIP to where the old RIP would have been in the old
instruction.  I think I couldn't do a calculation like this for
x86 because of the way that EIP was retrieved, and that I would
need to know the context of an instruction (like that a call
instruction was calling a thunk, and therefore the return value
was going to be used to determine EIP), whereas here we can know
from just one instruction that we are getting RIP.