Fixed instrumentation, refactored renabler.py, added global alloc
Instrumentation was broken since the major refactoring I did.  Now
the callback is actually assigned and instruments binaries similar
to how it did before.

Renabler.py is still not in the form I would like it for release,
because it is still an unpolished pile of functions (especially
the main driving function, now called rewrite instead of renable).
However, it is starting to approach a better state.  I eventually
want to merge the application arguments update I added on the x64
branch, but for now all scripts that call the tool from the
terminal use the old flag style, so I'll hold off changing that
yet.

I also added the ability to allocate an arbitrary amount of space
for global variables for use by the user when instrumenting
binaries.  For now, it allocates the space on the end of the
global lookup/mapping space, which is placed in files even when
they don't actually need it (i.e., when rewriting only the main
binary, it still puts the global mapping there). This could be
optimized out to only include the allocated globals, and
leave out the rest when rewriting only the main binary.  However,
doing that is fairly low priority.  I added this feature so
that I could store a global 64-bit counter instead of misusing
a different variable for that purpose.

Now the icount.py script counts instructions a bit differently, but
it now does it almost identically to how Pin does (when inserting
a counter increment before every instruction).  My approach saves
flags for every added instrumentation, while Pin can optimize the
obviously unneeded flag saving out.  Also, I don't know what will
happen for multithreaded applications, but the globals are not
thread safe, so I suspect the counter could be corrupted in
multithreaded applications.  This is something I'm not going to
worry about for now...