Finished initial changes to x64_runtime.py
Now, in theory, the rewriter should now be able to rewrite 64-bit
binaries with only the --execonly flag.  Unfortunately, this can't
be tested just yet because ELFManip doesn't actually support
64-bit binaries!  I think it mostly has to do with how the headers
are written out using struct.pack.  The problem is that even though
the entries are read in without assuming a particular size, a
32-bit binary is assumed when writing the entries.  Whether fixing
just this will be enough to make things work is unknown, but I have
the feeling there will be a few other problems before things work.