Got x64 support working for simple test program with --execonly
There were some things that still needed fixing with the PLT and
the lookup function, but after a bit of time tracking those
problems down, everything seems to be working for a simple program.

However, the program only works for the --execonly flag right now,
and I still need to check with bigger applications.