Add support for --nopic flag for x64
The fix was actually quite simple, since in supporting x64 I get
PIC support for free (I had to handle all references to rip, so
I already handle what x64 uses for PIC).  I just had to remove
references to the thunk, which is not used in x64.

I also made a minor fix for calling the lookup function for
callbacks.  It should have called the secondary lookup function,
but it was instead calling the regular lookup function, which
rewrites return addresses if they are outside the main binary.
We want to only lookup each address for the callbacks, and they
are not on the stack, so the secondary lookup is the right choice.