Changed popgm code to be platform specific
I am preparing to write the popgm code for x64, but I realized I
needed to make some modifications first to allow for different
versions of popgm for different architectures.  I also repurposed
the Makefile (which compiled a bunch of test programs that are not
needed now, and should probably be removed or moved to a test
directory) to automate the process of building popgm so that now
simply running make will build both x86_popgm and x64_popgm.  I
tested x86_popgm, and everything still works as expected for x86.

The x64 popgm code does not currently do anything useful.  I simply
made sure it compiled, but it is not functional.