Wrote all-binary x64 lookup, started fixing .so rip replacement
The x64 lookup function for rewriting everything (main binary and
libraries), should now be done, but still needs testing.

I also have finished a prototype for replacing rip in instructions
in .so files (a bit confusing because it needs to be computed
as rip-relative from the new location), but it also needs testing.