Continued work on replacing rip references in .so files
Since so many instructions have to be modified (all instructions
that reference rip, which can be any instruction that can reference
memory, I think), we encounter the problem that there are all sorts
of instructions to deal with.  I added a blacklist of incompatible
instructions, but I think I won't need it because of a side effect
of solving another problem: the rewriting became unbearably slow
again.

To solve the slowdown for rewriting in the past, I addressed the
primary bottleneck: the assembler.  Rewriting all these new
instructions brings the assembler bottleneck back.  The issue is
that the assembler I use currently, pwntools, is very slow, so
basically anything I can do to avoid calling it is faster.

I noticed that all the instructions that reference rip seem to
follow a pattern (supported by information I have read about
rip-relative addressing): there is ALWAYS a 4-byte offset at the
end of the instruction for the offset from rip, even if the offset
is zero (in which case, the last 4 bytes are zero).  Therefore, I
can simply take the bytes of the disassembled instruction and
replace the last 4 bytes with the new offset.  No need to call the
assembler at all!  This seems to work, and while I need to test
whether I am computing the new offset correctly, the data seems
correct.

However, I still cannot successfully rewrite libc, because I have
encountered another problem: if an offset is almost at the maximum
amount (high enough that only a slightly larger offset would
overflow 4 bytes), then the new offset may indeed not fit into a
4-byte offset!  I found an example with this instruction:
inc dword ptr [rip - 0x7ffb7378]
The rewritten instruction with the new offset is:
inc dword ptr [rip - 0x805955da]
This is outside the range supported by 4 bytes
(-0x80000000 to 0x7fffffff)!  Does this mean that this instruction
cannot be rewritten directly?  I think so.  Obviously it could be
changed to a series of instructions, but doing this generically
will probably be complex.  However, I do not believe the original
instruction was valid in this case; the address would be out of the
memory range of the loaded .so file, and I believe that a case in
which a real instruction has such a high offset will be very
uncommon.  Therefore, my plan is to put a placeholder offset when
the value is out of range, for now.